requirements.txt
google_streetview
youtube_transcript_api
openai
bs4
pytube
Pillow
retry
selenium